Added ClientsArray class and access from ServerThreads
All new ServerThreads can be added to the ClientsArray, where they are
stored for their lifetime. Each ServerThread has an incremental ID which
is used to identify it in the array when it is time to remove it. When
the ServerThread is disconnected from the ChatClient it throws an error
which when caught calls for itself to be removed from the ClientsArray.